Delete passenger API fixed

+++ src/main/java/edu/sjsu/services/PassengerService.java
@@ -41,28 +41,24 @@ public class PassengerService {
 	}
 	
 	public boolean deletePassenger(String id) {
-		try {
-			Passenger delPassenger = passengerDAO.getPassenger(id);
-			List<Reservation> reservations = reservationDAO.getReservations(id);
-			if(reservations.isEmpty())
-				System.out.println("EMPTYTPTTTYTYTYTYTYT");
-			else {
-				for(Reservation reservation : reservations)
-					System.out.println(reservation.getOrderNumber());
-			}
-			for(Reservation reservation : reservations) {
-				System.out.println("---->hhhh");
-				List<Flight> flights = reservation.getFlights();
-				for(Flight flight : flights) {
-					System.out.println("---->kkkk");
-					flight.setSeatsLeft(flight.getSeatsLeft() + 1);
-					flightDAO.createFlight(flight);
+		Passenger delPassenger = passengerDAO.getPassenger(id);
+		if(delPassenger != null) {
+			System.out.println("Passenger got");
+			List<Reservation> reservations = reservationDAO.getReservations(delPassenger);
+			if(reservations != null) {
+				for(Reservation reservation : reservations) {
+					List<Flight> flights = reservation.getFlights();
+					for(Flight flight : flights) {
+						flight.setSeatsLeft(flight.getSeatsLeft() + 1);
+						flight.getPassengers().remove(delPassenger);
+						flightDAO.createFlight(flight);
+					}
+					reservationDAO.deleteReservation(reservation);
 				}
-				reservationDAO.deleteReservation(reservation);
 			}
 			passengerDAO.deletePassenger(delPassenger);
 			return true;
-		} catch (Exception ex) {
+		} else {
 			return false;
 		}
 	}
